♻️ Extract form helpers to a reusable module
